requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.34.14
botocore==1.34.14
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.2.0
crispy-bootstrap4==2023.1
cryptography==41.0.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.8.0
dj-database-url==2.1.0
Django==4.2.5
django-allauth==0.56.1
django-crispy-forms==2.1
django-storages==1.14.2
fonttools==4.47.0
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-speech==2.23.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
install==1.3.5
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
openai==1.6.0
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
pocketsphinx==5.0.2
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.1
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.0.4
webrtcvad==2.0.10
yarl==1.9.4
